Refactor in preparation for ICMP scanning

+++ src/scan.h
@@ -14,7 +14,7 @@ struct ports;
 #define FIRST_SEQNUM 0xf0000000
 
 void scan_set_general(const struct ports *ports, int max_rate, int show_closed, int banners);
-void scan_set_network(const uint8_t *source_addr, int source_port, int ip_type);
+void scan_set_network(const uint8_t *source_addr, int source_port, uint8_t ip_type);
 void scan_set_output(FILE *outfile, const struct outputdef *outdef);
 int scan_main(const char *interface, int quiet);
